% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_app_intro_server.R
\name{app_intro}
\alias{app_intro}
\title{Shiny module server for the sse introduction page}
\usage{
app_intro(input, output, session)
}
\description{
Shiny module server for the sse introduction page
}
\seealso{
\code{\link{app_intro_ui}}
}
