% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_calc_prop.R
\name{calc_prop}
\alias{calc_prop}
\title{Evaluates power functions for power.prop.test}
\usage{
calc_prop(n, theta, xi, alpha, save_path)
}
\arguments{
\item{n}{numeric vector containing possible range of sample sizes}

\item{theta}{numeric vector containing possible range of effect sizes}

\item{xi}{numeric vector containing possible range of proportions}

\item{alpha}{numeric vector containing possible range of significance levels}

\item{save_path}{string inidicating where the function output should be saved}
}
\value{
list of calculated powCalc objects
}
\description{
For all pre-set combinations of n, theta, xi and alpha, this function calculates a list of powCalc
objects and saves the list.
}
\details{
Input parameters are optional. Defaults:
\itemize{
\item n <- seq(from = 3, to = 10000, by = 5)
\item theta <- c(seq(from = 0, to = 0.15, by = 0.0125), seq(from = 0.2, to = 0.5, by = 0.05))
\item xi <- seq(from = 0.05, to = 1, by = 0.05)
\item alpha <- c(0.01, 0.025, 0.05)
\item save_path <- file.path("inst", "extdata", "calc_prop_test.rds")
}
}
\seealso{
\code{\link{calc_t}}
}
