% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_calc_t.R
\name{calc_t}
\alias{calc_t}
\title{Evaluates power functions for power.t.test}
\usage{
calc_t(n, theta, xi, alpha, type, save_path)
}
\arguments{
\item{n}{numeric vector containing possible range of sample sizes}

\item{theta}{numeric vector containing possible range of effect sizes}

\item{xi}{numeric vector containing possible range of standard deviations}

\item{alpha}{numeric vector containing possible range of significance levels}

\item{type}{string of value "two.sample", "one.sample", or "paired"}

\item{save_path}{string inidicating where the function output should be saved}
}
\value{
list of calculated powCalc objects
}
\description{
For all pre-set combinations of n, theta, xi and alpha, this function calculates a list of powCalc
objects and saves them.
}
\details{
Input parameters are optional. Defaults:
\itemize{
\item n <- seq(from = 5, to = 2000, by = 5)
\item theta <- seq(from = 1, to = 50, by = 1)
\item xi <- seq(from = 1, to = 80, by = 1)
\item alpha <- c(0.01, 0.025, 0.05)
\item save_path <- file.path("inst", "extdata", "calc_t_test.rds")
}
}
