% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dash_ui.R
\name{dashUI}
\alias{dashUI}
\title{Shiny module UI function for the sse introduction page}
\usage{
dashUI(id, label)
}
\description{
Shiny module UI function for the sse introduction page
}
\seealso{
\code{\link{dash}}
}
