% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_calc_prop.R
\name{fun_calc_prop}
\alias{fun_calc_prop}
\title{Evaluates power -function for power.prop.test}
\usage{
fun_calc_prop()
}
\value{
list of calculated powPar objects
}
\description{
For all pre-set combinations of n, theta, xi and alpha, this function calculates a list of powPar (package: sse)
objects and saves them at a pre-set location in inst/extdata. The function uses the stats::power.prop.test for
this purpose. Make sure to re-build the package after running the function, in order to make it available
with package installation.
}
