% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_calc_t.R
\name{fun_calc_t}
\alias{fun_calc_t}
\title{Evaluates power -function for power.t.test}
\usage{
fun_calc_t(type = "two.sided")
}
\arguments{
\item{type}{character string specifying the type of t test. options: "two.sample", "one.sample", "paired"}
}
\value{
list of calculated powPar objects
}
\description{
For all pre-set combinations of n, theta, xi and alpha, this function calculates a list of powPar (package: sse)
objects and saves them at a pre-set location in inst/extdata. The function uses the stats::power.t.test for
this purpose. Make sure to re-build the package after running the function, in order to make it available
}
