% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_get_header_mint.R
\name{get_header_mint}
\alias{get_header_mint}
\title{Shinydashboard two-color custom mint header}
\usage{
get_header_mint(header_title)
}
\arguments{
\item{header_title}{string containing a header title}
}
\value{
shiny.tag list with a custom header
}
\description{
Shinydashboard two-color custom mint header
}
