% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_intro_server.R
\name{intro}
\alias{intro}
\title{Shiny module server function for the sse introduction page}
\usage{
intro(input, output, session)
}
\description{
Shiny module server function for the sse introduction page
}
\seealso{
\code{\link{introUI}}
}
