% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_intro_ui.R
\name{introUI}
\alias{introUI}
\title{Shiny module UI function for the sse introduction page}
\usage{
introUI(id, label)
}
\description{
Shiny module UI function for the sse introduction page
}
\seealso{
\code{\link{intro}}
}
