% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{run_shiny}
\alias{run_shiny}
\title{Runs the shinysse Shiny application}
\usage{
run_shiny()
}
\description{
You can run the web application by alling the run_shiny() function.
}
\seealso{
\code{\link{app}},\code{\link{app_ui}}
}
