% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sse_intro_server.R
\name{sse_intro}
\alias{sse_intro}
\title{Shiny module server function for the app introduction page}
\usage{
sse_intro(input, output, session)
}
\description{
Shiny module server function for the app introduction page
}
\seealso{
\code{\link{sse_intro_ui}}
}
