% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sse_intro_ui.R
\name{sse_intro_ui}
\alias{sse_intro_ui}
\title{Shiny module UI function for the app introduction page}
\usage{
sse_intro_ui(id, label)
}
\description{
Shiny module UI function for the app introduction page
}
\seealso{
\code{\link{sse_intro}}
}
