% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sse_prop_server.R
\name{sse_prop}
\alias{sse_prop}
\title{Shiny module server function for the display of sse for test for proportions}
\usage{
sse_prop(input, output, session, calcs)
}
\description{
Plots results of the power.prop.test for the two-sample test for proportions.
}
\seealso{
\code{\link{sse_prop_ui}}
}
