% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sse_t_test_server.R
\name{sse_t_test}
\alias{sse_t_test}
\title{Shiny module server function for the sse power.t.test module}
\usage{
sse_t_test(input, output, session, calcs_twosample, calcs_paired)
}
\description{
This shiny server function to be used for sse_t_test_ui().
}
\seealso{
\code{\link{sse_t_test_ui}}
}
