% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sse_t_test_ui.R
\name{sse_t_test_ui}
\alias{sse_t_test_ui}
\title{Shiny module UI function for the sse power.t.test module}
\usage{
sse_t_test_ui(id, label)
}
\description{
This shiny UI function to be used for sse_t_test() server.
}
\seealso{
\code{\link{sse_t_test}}
}
