% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sse_ttest_ui.R
\name{sse_ttest_ui}
\alias{sse_ttest_ui}
\title{Shiny module UI function for the display of sse for t-test}
\usage{
sse_ttest_ui(id, label)
}
\description{
Plots results of the power.t.test for the two-sample or paired t-test.
}
\seealso{
\code{\link{sse_ttest}}
}
