% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sseBasic_server.R
\name{sseBasic}
\alias{sseBasic}
\title{Shiny module server function for the sse power.t.test module}
\usage{
sseBasic(input, output, session, calcs_twosample, calcs_paired)
}
\description{
This shiny server function to be used for sseBasicUI().
}
\seealso{
\code{\link{sseBasicUI}}
}
