% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sseBasic_ui.R
\name{sseBasicUI}
\alias{sseBasicUI}
\title{Shiny module UI function for the sse power.t.test module}
\usage{
sseBasicUI(id, label)
}
\description{
This shiny UI function to be used for sseBasic() server.
}
\seealso{
\code{\link{sseBasic}}
}
