% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ssePlus_server.R
\name{ssePlus}
\alias{ssePlus}
\title{Shiny module server function for the sse power.t.test module with user input}
\usage{
ssePlus(input, output, session, calcs)
}
\description{
This shiny server function to be used for ssePlusUI().
}
\seealso{
\code{\link{ssePlusUI}}
}
