% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ssePlus_ui.R
\name{ssePlusUI}
\alias{ssePlusUI}
\title{Shiny module UI function for the sse power.t.test module with user input}
\usage{
ssePlusUI(id, label)
}
\description{
This UI gives the user an opportunity to enter two group means as numeric inputs, in addition to standard deviation,
etc.
}
\seealso{
\code{\link{ssePlus}}
}
