% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sseProp_server.R
\name{sseProp}
\alias{sseProp}
\title{Shiny module server function for the sse power.prop.test module}
\usage{
sseProp(input, output, session, calcs)
}
\description{
Shiny module server function for the sse power.prop.test module
}
\seealso{
\code{\link{ssePropUI}}
}
