% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sseProp_ui.R
\name{ssePropUI}
\alias{ssePropUI}
\title{Shiny module UI function for the sse power.prop.test module}
\usage{
ssePropUI(id, label)
}
\description{
Shiny module UI function for the sse power.prop.test module
}
\seealso{
\code{\link{sseProp}}
}
