% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_intro_server.R
\name{app_intro}
\alias{app_intro}
\title{Shiny module server function for the app introduction page}
\usage{
app_intro(input, output, session)
}
\description{
Shiny module server function for the app introduction page
}
\seealso{
\code{\link{app_intro}}
}
