% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_run_shiny.R
\name{run_shiny}
\alias{run_shiny}
\title{Launches the shinysse Shiny web app}
\usage{
run_shiny()
}
\description{
You can run the web application by calling the run_shiny() function.
}
\seealso{
\code{\link{app_server}}, \code{\link{app_ui}}
}
