% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sse_prop_server.R
\name{sse_prop}
\alias{sse_prop}
\title{Shiny module server function for the sse power.prop.test module}
\usage{
sse_prop(input, output, session, calcs)
}
\description{
Shiny module server function for the sse power.prop.test module
}
\seealso{
\code{\link{sse_prop_ui}}
}
