% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sse_prop_ui.R
\name{sse_prop_ui}
\alias{sse_prop_ui}
\title{Shiny module UI function for the sse power.prop.test module}
\usage{
sse_prop_ui(id, label)
}
\description{
Shiny module UI function for the sse power.prop.test module
}
\seealso{
\code{\link{sse_prop}}
}
