% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sse_t_test_server.R
\name{sse_ttest}
\alias{sse_ttest}
\title{Shiny module server function for the sse power.t.test module}
\usage{
sse_ttest(input, output, session, calcs_twosample, calcs_paired)
}
\description{
This shiny server function to be used for sse_ttest_ui().
}
\seealso{
\code{\link{sse_ttest_ui}}
}
