% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sse_ttest_ui.R
\name{sse_ttest_plus_ui}
\alias{sse_ttest_plus_ui}
\title{Shiny module UI function for the sse power.t.test module with user input}
\usage{
sse_ttest_plus_ui(id, label)
}
\description{
This UI gives the user an opportunity to enter two group means as numeric inputs, in addition to standard deviation,
etc.
}
\seealso{
\code{\link{sse_ttest_plus}}
}
