% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sse_t_test_ui.R
\name{sse_ttest_ui}
\alias{sse_ttest_ui}
\title{Shiny module UI function for the sse power.t.test module}
\usage{
sse_ttest_ui(id, label)
}
\description{
This shiny UI function to be used for sse_ttest() server.
}
\seealso{
\code{\link{sse_ttest}}
}
